+++ devops/build.sh
@@ -19,13 +19,13 @@ echo "Tag: $FIN_TAG_NAME"
 echo "SHA: $FIN_SERVER_REPO_HASH"
 
 echo -e "\nBuilding images:"
-echo    "Fcrepo        : $FCREPO_IMAGE_NAME:$APP_TAG"
-echo    "Postgres      : $POSTGRES_IMAGE_NAME:$APP_TAG"
-echo    "Apache LB     : $LB_IMAGE_NAME:$APP_TAG"
-echo    "Base Service  : $SERVER_IMAGE_NAME:$APP_TAG"
-echo    "ElasticSearch : $ELASTIC_SEARCH_IMAGE_NAME:$APP_TAG"
-echo    "PG Rest       : $PGREST_IMAGE_NAME:$APP_TAG"
-echo -e "Init          : $INIT_IMAGE_NAME:$APP_TAG\n"
+echo    "Fcrepo        : $FCREPO_IMAGE_NAME:$APP_TAG and :$DOCKER_CACHE_TAG"
+echo    "Postgres      : $POSTGRES_IMAGE_NAME:$APP_TAG and :$DOCKER_CACHE_TAG"
+echo    "Apache LB     : $LB_IMAGE_NAME:$APP_TAG and :$DOCKER_CACHE_TAG"
+echo    "Base Service  : $SERVER_IMAGE_NAME:$APP_TAG and :$DOCKER_CACHE_TAG"
+echo    "ElasticSearch : $ELASTIC_SEARCH_IMAGE_NAME:$APP_TAG and :$DOCKER_CACHE_TAG"
+echo    "PG Rest       : $PGREST_IMAGE_NAME:$APP_TAG and :$DOCKER_CACHE_TAG"
+echo -e "Init          : $INIT_IMAGE_NAME:$APP_TAG and :$DOCKER_CACHE_TAG\n"
 
 # Core Server - fcrepo
 docker build \
@@ -36,6 +36,7 @@ docker build \
   -t $FCREPO_IMAGE_NAME:$APP_TAG \
   --cache-from $FCREPO_IMAGE_NAME:$DOCKER_CACHE_TAG \
   services/fcrepo
+docker tag $FCREPO_IMAGE_NAME:$APP_TAG $FCREPO_IMAGE_NAME:$DOCKER_CACHE_TAG
 
 # Core Server - postgres
 docker build \
@@ -46,6 +47,7 @@ docker build \
   -t $POSTGRES_IMAGE_NAME:$APP_TAG \
   --cache-from $POSTGRES_IMAGE_NAME:$DOCKER_CACHE_TAG \
   services/postgres
+docker tag $POSTGRES_IMAGE_NAME:$APP_TAG $POSTGRES_IMAGE_NAME:$DOCKER_CACHE_TAG
 
 # Core Server - apache lb
 docker build \
@@ -56,6 +58,7 @@ docker build \
   -t $LB_IMAGE_NAME:$APP_TAG \
   --cache-from $LB_IMAGE_NAME:$DOCKER_CACHE_TAG \
   services/load-balancer
+docker tag $LB_IMAGE_NAME:$APP_TAG $LB_IMAGE_NAME:$DOCKER_CACHE_TAG
 
 # Core Server - server
 docker build \
@@ -67,6 +70,7 @@ docker build \
   --cache-from $SERVER_IMAGE_NAME:$DOCKER_CACHE_TAG \
   -f services/fin/Dockerfile \
   .
+docker tag $SERVER_IMAGE_NAME:$APP_TAG $SERVER_IMAGE_NAME:$DOCKER_CACHE_TAG
 
 # Core Server - elastic search
 docker build \
@@ -77,6 +81,7 @@ docker build \
   -t $ELASTIC_SEARCH_IMAGE_NAME:$APP_TAG \
   --cache-from $ELASTIC_SEARCH_IMAGE_NAME:$DOCKER_CACHE_TAG \
   services/elastic-search
+docker tag $ELASTIC_SEARCH_IMAGE_NAME:$APP_TAG $ELASTIC_SEARCH_IMAGE_NAME:$DOCKER_CACHE_TAG
 
 # Core - Init services
 docker build \
@@ -88,6 +93,7 @@ docker build \
   -t $INIT_IMAGE_NAME:$APP_TAG \
   --cache-from $INIT_IMAGE_NAME:$DOCKER_CACHE_TAG \
   services/init
+docker tag $INIT_IMAGE_NAME:$APP_TAG $INIT_IMAGE_NAME:$DOCKER_CACHE_TAG
 
 # Core - PG REST
 docker build \
@@ -98,3 +104,4 @@ docker build \
   -t $PGREST_IMAGE_NAME:$APP_TAG \
   --cache-from $PGREST_IMAGE_NAME:$DOCKER_CACHE_TAG \
   services/pg-rest
+docker tag $PGREST_IMAGE_NAME:$APP_TAG $PGREST_IMAGE_NAME:$DOCKER_CACHE_TAG