+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2.0.3
+
+ - Disabling autoversioning of ocfl.  `fcrepo.autoversioning.enabled=false`
+ - Fix for exposing pg uuid to public schema
+ 
 
 # 2.0.2